[Silver V] Title: 타일 장식물, Time: 108 ms, Memory: 14188 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13301. 타일 장식물/타일 장식물.java
@@ -6,7 +6,7 @@ public class Main
 	public static void main(String[] args) throws Exception{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int n = Integer.parseInt(br.readLine());
-		int[] dp = new int[n + 1];
+		long[] dp = new long[n + 1];
 		dp[0] = 1;
 		dp[1] = 1;
 		for(int i = 2; i <= n; i++){
